requirements.txt
flask
google-cloud